Add linter
and fix some lint errors

NB. Depends on an old version of the js-sdk because the js-sdk
doesn't export our linting rules anymore (react-sdk checks the
develop branch out so it gets the very latest rules). We think
the best fix for this is to split our eslint rules out into a
separate repo. This isn't a mission I want to embark on now, so
pulling in the old js-sdk to have linting while I write this.