Don't pull either the artifacts or the deb database
This is now unnecessary because the build box is the only box writing
artifacts to the build server, so we can just consider whatever's on
the build box as authoritive and clobber whatever's on the server.

The reprepro database simlarly no longer needs to be synced to the server,
since there's only one thing writing to it: we can just keep the
database on the build box.

This should fix a whole multitude of failure modes where changes on
the build box get clobbered when it syncs from the server.